Moved binding to base fragment.
